requirements.txt
absl-py==2.1.0
appdirs==1.4.4
argcomplete==3.5.3
astunparse==1.6.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imageio==2.35.1
importlib_resources==6.5.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
Pygments==2.18.0
pymodbus==3.6.9
pyModbusTCP==0.2.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-barcode==0.15.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-escpos==3.1
pytz==2024.1
pywin32==308
PyYAML==6.0.1
qrcode==8.0
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sympy==1.13.0
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.55
ultralytics-thop==2.0.0
ur-rtde==1.5.7
urllib3==2.2.2
waitress==3.0.2
Werkzeug==3.0.3
wrapt==1.16.0
